requirements.txt
kivy
git+https://github.com/dolfies/discord.py-self.git
cryptography
Pillow
requests
pyvips
BeautifulSoup4
g4f
Brotli